+++ src/babel/__tests__/examples/anonymousElements.js
@@ -1,0 +1,4 @@
+import * as zacs from '@nozbe/zacs'
+
+const a = <zacs.view />
+const b = <zacs.text zacs:style={{ width: 100 }} />