+++ src/babel/nonZacsElements.js
@@ -1,8 +1,10 @@
 const { types: t } = require('@babel/core')
-const { getPlatform } = require('./state')
-const { jsxAttr, jsxFindNamespacedAttrPath } = require('./jsxUtils')
+const { getPlatform, isProduction } = require('./state')
+const { getElementName } = require('./declarations')
+const { jsxAttr, jsxFindNamespacedAttrPath, jsxRenameElement } = require('./jsxUtils')
 const { mergeObjects, concatArraysOfObjects } = require('./astUtils')
 const { resolveZacsStyleAttr } = require('./resolveStyle')
+const { setUsesRN } = require('./imports')
 
 function mergeStyles(platform, maybeZacsStyleExpr, maybeZacsInheritStyleExpr) {
   // NOTE: zacs:style comes before zacs:inherit
@@ -15,6 +17,52 @@ function mergeStyles(platform, maybeZacsStyleExpr, maybeZacsInheritStyleExpr) {
   }
 
   throw new Error('Unknown platform')
+}
+
+function transformAnonymousZacsElement(path, state) {
+  const platform = getPlatform(state)
+
+  const { node } = path
+  const { openingElement } = node
+
+  if (
+    !(
+      t.isJSXMemberExpression(openingElement.name) &&
+      t.isJSXIdentifier(openingElement.name.object, { name: 'zacs' })
+    )
+  ) {
+    return
+  }
+
+  if (
+    !(t.isJSXIdentifier(openingElement.name.property),
+    ['view', 'text'].includes(openingElement.name.property.name))
+  ) {
+    throw path
+      .get('openingElement.name.property')
+      .buildCodeFrameError(
+        `zacs.${openingElement.name.property.name} is not a valid element. Did you mean <zacs.view /> or <zacs.text /> ?`,
+      )
+  }
+
+  const elementType = openingElement.name.property.name
+  const originalName = `zacs.${elementType}`
+
+  const [elementName, isBuiltin] = getElementName(platform, path, originalName, elementType)
+
+  if (platform === 'native') {
+    setUsesRN(state, elementName)
+  }
+
+  // replace component
+  jsxRenameElement(node, elementName, isBuiltin)
+
+  // add debug info
+  if (!isProduction(state)) {
+    openingElement.attributes.unshift(
+      t.jSXAttribute(t.jSXIdentifier('__zacs_original_name'), t.stringLiteral(originalName)),
+    )
+  }
 }
 
 function transformZacsAttributesOnNonZacsElement(path, state) {
@@ -65,4 +113,4 @@ function transformZacsAttributesOnNonZacsElement(path, state) {
     .concat(addedAttrs)
 }
 
-module.exports = { transformZacsAttributesOnNonZacsElement }
+module.exports = { transformAnonymousZacsElement, transformZacsAttributesOnNonZacsElement }