+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - `'.foo, .bar': {}` CSS-only stylesets
 - `border: [1, 'red']` shorthand
 - You can now use shorthands and `web/native: {}` scopes inside `zacs:style={}` attribute. Note that this only works if an object literal is passed. You can transform any such style object with `zacs._experimental_resolve({ ... })`
+- `<zacs.view />` and `<zacs.text />` anonymous ZACS elements
 
 **Performance**